Rebase fixes
